+++ apps/public-docsite/webpack.serve.config.js
@@ -1,6 +1,7 @@
 // @ts-check
 
 const path = require('path');
+const webpack = require('webpack');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const resources = require('../../scripts/webpack/webpack-resources');
 const getResolveAlias = require('../../scripts/webpack/getResolveAlias');
@@ -17,10 +18,11 @@ const outDir = 'dist';
 module.exports = [
   // Copy index.html and generate bootstrap script
   getLoadSiteConfig({
-    libraryName: '@fluentui/react',
+    libraryPath: path.dirname(require.resolve('@fluentui/react/package.json')),
     outDir: path.join(__dirname, outDir),
     isProduction: false,
     CopyWebpackPlugin,
+    webpack,
   }),
   // Rest of site
   resources.createServeConfig(