+++ packages/react-menu/src/components/MenuItem/MenuItem.types.ts
@@ -13,9 +13,25 @@ export interface MenuItemProps extends ComponentProps, React.HTMLAttributes<HTML
   submenuIndicator?: ShorthandProps<React.HTMLAttributes<HTMLElement>>;
 
   /**
+   * Component children are placed in this slot
+   * Avoid using the `children` property in this slot in favour of Component children whenever possible
+   */
+  content?: ShorthandProps<React.HTMLAttributes<HTMLElement>>;
+
+  /**
+   * Secondary content rendered opposite the primary content (e.g Shortcut text)
+   */
+  secondaryContent?: ShorthandProps<React.HTMLAttributes<HTMLElement>>;
+
+  /**
    * If the menu item is a trigger for a submenu
    */
   hasSubmenu?: boolean;
+
+  /**
+   * Applies disabled styles to menu item but remains focusable
+   */
+  disabled?: boolean;
 }
 
 export interface MenuItemState extends MenuItemProps {
@@ -23,6 +39,7 @@ export interface MenuItemState extends MenuItemProps {
    * Ref to the root slot
    */
   ref: React.MutableRefObject<HTMLElement>;
+
   /**
    * Icon slot when processed by internal state
    */
@@ -32,4 +49,14 @@ export interface MenuItemState extends MenuItemProps {
    * Icon slot that shows the indicator for a submenu
    */
   submenuIndicator?: ObjectShorthandProps<React.HTMLAttributes<HTMLElement>>;
+
+  /**
+   * Slot for the component children, avoid in favour of children and classnames for customization
+   */
+  content: ObjectShorthandProps<React.HTMLAttributes<HTMLElement>>;
+
+  /**
+   * Secondary content rendered opposite the primary content (e.g Shortcut text)
+   */
+  secondaryContent: ObjectShorthandProps<React.HTMLAttributes<HTMLElement>>;
 }