+++ packages/react-menu/src/components/MenuItem/useMenuItem.tsx
@@ -15,7 +15,7 @@ import { ChevronRightIcon } from '../../utils/DefaultIcons';
  * Consts listing which props are shorthand props.
  */
 // TODO introduce content slot for styling
-export const menuItemShorthandProps = ['icon', 'submenuIndicator'] as const;
+export const menuItemShorthandProps = ['icon', 'submenuIndicator', 'content', 'secondaryContent'] as const;
 
 // eslint-disable-next-line deprecation/deprecation
 const mergeProps = makeMergePropsCompat<MenuItemState>({ deepMerge: menuItemShorthandProps });
@@ -35,22 +35,41 @@ export const useMenuItem = (
       ref: useMergedRefs(ref, React.useRef(null)),
       icon: { as: 'span' },
       submenuIndicator: { as: 'span', children: <ChevronRightIcon /> },
+      content: { as: 'span', children: props.children },
+      secondaryContent: { as: 'span' },
       role: 'menuitem',
       tabIndex: 0,
       hasSubmenu,
+      'aria-disabled': props.disabled,
     },
     defaultProps,
     resolveShorthandProps(props, menuItemShorthandProps),
   );
 
-  const { onKeyDown: onKeyDownOriginal } = state;
+  const { onClick: onClickOriginal, onKeyDown: onKeyDownOriginal } = state;
   state.onKeyDown = (e: React.KeyboardEvent<HTMLElement>) => {
     if (shouldPreventDefaultOnKeyDown(e)) {
+      if (state.disabled) {
+        e.preventDefault();
+        e.stopPropagation();
+        return;
+      }
+
       e.preventDefault();
       (e.target as HTMLElement)?.click();
     }
 
     onKeyDownOriginal?.(e);
+  };
+
+  state.onClick = (e: React.MouseEvent<HTMLElement>) => {
+    if (state.disabled) {
+      e.preventDefault();
+      e.stopPropagation();
+      return;
+    }
+
+    onClickOriginal?.(e);
   };
 
   const { onMouseEnter: onMouseEnterOriginal } = state;