+++ packages/react-menu/src/components/MenuItem/useMenuItemStyles.ts
@@ -5,7 +5,6 @@ const useStyles = makeStyles({
   root: theme => ({
     color: theme.alias.color.neutral.neutralForeground1,
     backgroundColor: theme.alias.color.neutral.neutralBackground1,
-    // TODO when introducing secondary text right padding might need to change
     paddingRight: '8px',
     paddingLeft: '12px',
     height: '32px',
@@ -24,6 +23,22 @@ const useStyles = makeStyles({
       color: theme.alias.color.neutral.neutralForeground2Hover,
     },
   }),
+  content: {
+    marginRight: '8px',
+    backgroundColor: 'transparent',
+    flexGrow: 1,
+  },
+  secondaryContent: theme => ({
+    color: theme.alias.color.neutral.neutralForeground3,
+    ':hover': {
+      color: theme.alias.color.neutral.neutralForeground3Hover,
+    },
+
+    ':focus': {
+      backgroundColor: theme.alias.color.neutral.neutralBackground1Hover,
+      color: theme.alias.color.neutral.neutralForeground3Hover,
+    },
+  }),
   icon: {
     width: '20px',
     height: '20px',
@@ -32,14 +47,30 @@ const useStyles = makeStyles({
   submenuIndicator: {
     width: '20px',
     height: '20px',
-    marginLeft: 'auto',
   },
+  disabled: theme => ({
+    backgroundColor: theme.alias.color.neutral.neutralBackgroundDisabled,
+    color: theme.alias.color.neutral.neutralForegroundDisabled,
+    ':hover': {
+      backgroundColor: theme.alias.color.neutral.neutralBackgroundDisabled,
+      color: theme.alias.color.neutral.neutralForegroundDisabled,
+    },
+
+    ':focus': {
+      backgroundColor: theme.alias.color.neutral.neutralBackgroundDisabled,
+      color: theme.alias.color.neutral.neutralForegroundDisabled,
+    },
+  }),
 });
 
 /** Applies style classnames to slots */
 export const useMenuItemStyles = (state: MenuItemState) => {
   const styles = useStyles();
-  state.className = ax(styles.root, state.className);
+  state.className = ax(styles.root, state.disabled && styles.disabled, state.className);
+  state.content.className = ax(styles.content, state.content.className);
+  if (state.secondaryContent) {
+    state.secondaryContent.className = ax(!state.disabled && styles.secondaryContent, state.secondaryContent.className);
+  }
 
   if (state.icon) {
     state.icon.className = ax(styles.icon, state.icon.className);