+++ packages/react-utilities/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './getNativeElementProps';
 export * from './omit';
 export * from './properties';
 export * from './shouldPreventDefaultOnKeyDown';
+export * from './isSSR';