+++ packages/react-utilities/src/utils/isSSR.ts
@@ -1,0 +1,6 @@
+/**
+ * Verifies if application is running in server side rendering or in the browser
+ */
+export function isSSR() {
+  return typeof window === 'undefined' || /ServerSideRendering/.test(window.navigator && window.navigator.userAgent);
+}