+++ packages/react-tabs/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @fluentui/react-tabs
 
-This log was last generated on Wed, 31 Mar 2021 00:53:43 GMT and should not be manually modified.
+This log was last generated on Tue, 06 Apr 2021 07:34:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.0-beta.76](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.76)
+
+Tue, 06 Apr 2021 07:34:10 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-tabs_v1.0.0-beta.75..@fluentui/react-tabs_v1.0.0-beta.76)
+
+### Patches
+
+- Bump @fluentui/react to v8.9.0 ([PR #17698](https://github.com/microsoft/fluentui/pull/17698) by tristan.watanabe@gmail.com)
+
+## [1.0.0-beta.75](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.75)
+
+Thu, 01 Apr 2021 07:33:24 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-tabs_v1.0.0-beta.74..@fluentui/react-tabs_v1.0.0-beta.75)
+
+### Patches
+
+- Bump @fluentui/react to v8.8.0 ([PR #17632](https://github.com/microsoft/fluentui/pull/17632) by miclo@microsoft.com)
 
 ## [1.0.0-beta.74](https://github.com/microsoft/fluentui/tree/@fluentui/react-tabs_v1.0.0-beta.74)